Refactor to put nmos::connection_activation_thread (back) into nmos module
(cherry picked from commit 87605551e98f90d789f658078f7c6d3e8d1b86e1)
