+++ Development/nmos/connection_api.cpp
@@ -560,7 +560,7 @@ namespace nmos
         // also be set back to null to unblock concurrent patch operations (nmos::set_connection_resource_not_pending does both).
         //
         // This obviously requires co-operation from other threads that are manipulating these connection resources.
-        // nmos-cpp-node/node_implementation.cpp currently serves as an example of how to handle sender/receiver activations.
+        // nmos::connection_activation_thread can usually be used to schedule sender/receiver activations.
         //
         // By the time we reacquire the model lock anything may have happened, but we can identify with the above whether to send
         // a success response or an error, and in the success case, release the 'per-resource lock' by updating the staged
@@ -731,7 +731,7 @@ namespace nmos
 
         void notify_connection_resource_patch(const nmos::node_model& model, slog::base_gate& gate)
         {
-            slog::log<slog::severities::too_much_info>(gate, SLOG_FLF) << "Notifying connection thread";
+            slog::log<slog::severities::too_much_info>(gate, SLOG_FLF) << "Notifying connection activation thread";
 
             model.notify();
         }
